refactor: centralize lifecycle listener default
Deduplicate Purchases lifecycle listener forwarding by using a shared default provider across flavor-specific Purchases entrypoints.
