+++ purchases/src/customEntitlementComputation/kotlin/com/revenuecat/purchases/Purchases.kt
@@ -213,7 +213,7 @@ public class Purchases internal constructor(
     public companion object {
         @OptIn(InternalRevenueCatAPI::class)
         @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
-        internal var lifecycleListener: PurchasesLifecycleListener = PurchasesLifecycleEventBusListener
+        internal var lifecycleListener: PurchasesLifecycleListener = PurchasesLifecycleListeners.default()
 
         /**
          * DO NOT MODIFY. This is used internally by the Hybrid SDKs to indicate which platform is
@@ -389,17 +389,6 @@ public class Purchases internal constructor(
             PurchasesOrchestrator.canMakePayments(context, features, callback)
         }
     }
-
-@OptIn(InternalRevenueCatAPI::class)
-private object PurchasesLifecycleEventBusListener : PurchasesLifecycleListener {
-    override fun onPurchasesConfigured(purchases: Purchases) {
-        PurchasesLifecycleEventBus.onConfigured(purchases)
-    }
-
-    override fun onPurchasesClosed(purchases: Purchases) {
-        PurchasesLifecycleEventBus.onClosed(purchases)
-    }
-}
 
     // endregion
 }
